Revert "Fix: main_svc"
This reverts commit 1836c3571fbb1abb3e6f4125355866de53b2a05b.

+++ ewm-main-service/src/main/java/ru/practicum/main/model/Location.java
@@ -20,10 +20,10 @@ public class Location {
     private Long id;
 
     @Column(name = "lat")
-    private float lat;
+    private double lat;
 
     @Column(name = "lon")
-    private float lon;
+    private double lon;
 
     @Override
     public String toString() {
